[ES|QL] Move the sources response result to the types package
