+++ src/platform/packages/shared/kbn-esql-validation-autocomplete/src/shared/types.ts
@@ -6,11 +6,7 @@
  * your election, the "Elastic License 2.0", the "GNU Affero General Public
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
-import {
-  Location,
-  ESQLSourceResult,
-  ESQLFieldWithMetadata,
-} from '@kbn/esql-ast/src/commands_registry/types';
+import { Location, ESQLFieldWithMetadata } from '@kbn/esql-ast/src/commands_registry/types';
 import { PricingProduct } from '@kbn/core-pricing-common/src/types';
 import type {
   ESQLControlVariable,
@@ -19,6 +15,7 @@ import type {
   RecommendedField,
   InferenceEndpointsAutocompleteResult,
   ESQLLicenseResult,
+  ESQLSourceResult,
 } from '@kbn/esql-types';
 import type { InferenceTaskType } from '@elastic/elasticsearch/lib/api/types';
 /** @internal **/
